% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{addMutator}
\alias{addMutator}
\title{Adds a mutation operator to a collection.}
\usage{
addMutator(collection, fun, ...)
}
\arguments{
\item{collection}{[\code{tspgen_collection}]\cr
Collection of mutation operators.}

\item{fun}{[\code{character(1)}]\cr
Function name of mutation operator which should be added to collection
provided as a string.}

\item{...}{[any]\cr
Parameters for \code{fun}. See the documentation of the corresponding mutation
operator for details.}
}
\value{
[\code{tspgen_collection}] Augmented mutation operator collection.
}
\description{
Adds a mutation operator to a collection.
}
