% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{applyRandomMutation}
\alias{applyRandomMutation}
\title{Apply a random mutation from a collection.}
\usage{
applyRandomMutation(collection, coords, bound.handling = "boundary")
}
\arguments{
\item{collection}{[\code{tspgen_collection}]\cr
Collection of mutation operators.}

\item{coords}{[\code{matrix(n, 2)}]\cr
Coordinate matrix.}

\item{bound.handling}{[\code{character(1)}]\cr
  Occasionally during instance generation points are moved outside the boundaries
  of the point space. The parameter \code{bound.handling} determines how to deal
  with these points. Option \dQuote{uniform} places outliers uniform at random within
.  the boundaries while option \dQuote{boundary} places them on the corresponding
  violates boundary/boundaries.
  Default is \dQuote{uniform}.}
}
\value{
[\code{matrix(n, 2)}] Mutated coordinate matrix.
}
\description{
Apply a random mutation from a collection.
}
