% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.axisprojection.R
\name{doAxisProjectionMutation}
\alias{doAxisProjectionMutation}
\title{Axis projection mutation}
\usage{
doAxisProjectionMutation(coords, pm = 0.1, p.jitter = 0,
  jitter.sd = 0.1, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{pm}{[\code{numeric(1)}]\cr
Probability of node mutation. Note that each node is subject to mutation
independent of the the nodes.
Default is \eqn{0.1}, i.e., in expectation 10\% of the points are mutated.}

\item{p.jitter}{[\code{numeric(1)}]\cr
Probability to add Gaussian noise to mutated points.
Default is 0, i.e., no Gaussian noise at all.}

\item{jitter.sd}{[\code{numeric(1)}]\cr
Standard deviation for Gaussian noise.
Defaults to 0.1. Note that this parameter has an effect only if \code{p.jitter} is
greater zero.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
A subset \eqn{Q \subseteq P} (each point selected with independent
probability \code{pm}) is selected. Next, an axis \eqn{a \in \{1, 2\}} and a constant
\eqn{c \in [0, 1]} are sampled. Finally, the mutator sets \eqn{p_a = c} for all \eqn{p \in Q},
i.e., a projection on the constant vertical or horizontal line is performed.
Optionally, with probability \code{p.jitter} all projected points are subject to
Gaussian noise with standard deviation \code{jitter.sd} with respect to the \eqn{a}-th
coordinate: \eqn{p_a = p_a + N(0, jitter.sd)}.
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doClusterMutation}},
  \code{\link{doCompressionMutation}},
  \code{\link{doExpansionMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doImplosionMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doNormalMutation}},
  \code{\link{doRotationMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
