% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.cluster.R
\name{doClusterMutation}
\alias{doClusterMutation}
\title{Cluster mutation}
\usage{
doClusterMutation(coords, pm = 0.1, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{pm}{[\code{numeric(1)}]\cr
Probability of node mutation. Note that each node is subject to mutation
independent of the the nodes.
Default is \eqn{0.1}, i.e., in expectation 10\% of the points are mutated.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
A subset \eqn{Q \subseteq P} (each point selected with independent
probability \code{pm}) is replaced by points placed according to a random sampleof size \eqn{|Q|} of
a bi-variate Gaussian distribution with uniformly sampled center and standard
deviation up to 0.3.
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doAxisProjectionMutation}},
  \code{\link{doCompressionMutation}},
  \code{\link{doExpansionMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doImplosionMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doNormalMutation}},
  \code{\link{doRotationMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
