% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.compression.R
\name{doCompressionMutation}
\alias{doCompressionMutation}
\title{Compression mutation}
\usage{
doCompressionMutation(coords, min.eps = 0.1, max.eps = 0.3, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{min.eps}{[\code{numeric(1)}]\cr
Minimum value for sampled tube width.
Default is 0.1}

\item{max.eps}{[\code{numeric(1)}]\cr
Maximum value for the sampled tube with.
Default is 0.3.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
This operator is very similar to the expansion mutation (see
\code{\link{doExpansionMutation}}). Hence we refer the reader to the corresponding
documentation for details. The main difference is that instead of moving points
away from their orthogonal projection (expansion) points are attracted by their
orthogonal projections on the linear function.
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doAxisProjectionMutation}},
  \code{\link{doClusterMutation}},
  \code{\link{doExpansionMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doImplosionMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doNormalMutation}},
  \code{\link{doRotationMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
