% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.expansion.R
\name{doExpansionMutation}
\alias{doExpansionMutation}
\title{Expansion mutation}
\usage{
doExpansionMutation(coords, min.eps = 0.1, max.eps = 0.3, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{min.eps}{[\code{numeric(1)}]\cr
Minimum value for sampled tube width.
Default is 0.1}

\item{max.eps}{[\code{numeric(1)}]\cr
Maximum value for the sampled tube with.
Default is 0.3.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
This is a generalization of the linear projection mutation operator
(see \code{\link{doLinearProjectionMutation}}). The parameters \eqn{a} (intercept)
and \eqn{m} (slope) of a linear function \eqn{a + mx} are sampled (see documentation
of \code{\link{doLinearProjectionMutation}} for details. All points whose orthogonal distance
to the linear function is below the sampled tube with \eqn{\epsilon = U[min.eps, max.eps]}
are subject to mutation. This is achieved by moving the points away from their orthogonal
projections to distance \eqn{\epsilon + Exp(\lambda = 10)}.
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doAxisProjectionMutation}},
  \code{\link{doClusterMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doImplosionMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doNormalMutation}},
  \code{\link{doRotationMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
