% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.implosion.R
\name{doImplosionMutation}
\alias{doImplosionMutation}
\title{Implosion mutation}
\usage{
doImplosionMutation(coords, min.eps = 0.1, max.eps = 0.3, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{min.eps}{[\code{numeric(1)}]\cr
Minimum value for sampled implosion radius.
Defaults to 0.1}

\item{max.eps}{[\code{numeric(1)}]\cr
Maximum value for sampled implosion radius.
Defaults to 0.3.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
Implosion mutation
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doAxisProjectionMutation}},
  \code{\link{doClusterMutation}},
  \code{\link{doCompressionMutation}},
  \code{\link{doExpansionMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doNormalMutation}},
  \code{\link{doRotationMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
