% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.normal.R
\name{doNormalMutation}
\alias{doNormalMutation}
\title{Normal mutation}
\usage{
doNormalMutation(coords, pm = 0.1, sigma = 0.0025, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{pm}{[\code{numeric(1)}]\cr
Probability of node mutation. Note that each node is subject to mutation
independent of the the nodes.
Default is \eqn{0.1}, i.e., in expectation 10\% of the points are mutated.}

\item{sigma}{[\code{numeric(1)}]\cr
Standard deviation for normal random numbers generator.
Default is \eqn{0.0025}.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
Each point is subject to additive Gaussian noise with probability \code{pm}.
}
\references{
Mersmann, O., Bischl, B., Bossek, J., Trautmann, H., Wagner, M., & Neumann, F. (2012).
Local search and the traveling salesman problem: A feature-based characterization of
problem hardness. Lecture Notes in Computer Science (Including Subseries Lecture Notes
in Artificial Intelligence and Lecture Notes in Bioinformatics), 7219 LNCS, 115-129.
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doAxisProjectionMutation}},
  \code{\link{doClusterMutation}},
  \code{\link{doCompressionMutation}},
  \code{\link{doExpansionMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doImplosionMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doRotationMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
