% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutator.rotation.R
\name{doRotationMutation}
\alias{doRotationMutation}
\title{Rotation mutation}
\usage{
doRotationMutation(coords, pm = 0.1, ...)
}
\arguments{
\item{coords}{[\code{matrix}]\cr
An n times 2 matrix of point coordinates in \eqn{[0, 1]^2}.}

\item{pm}{[\code{numeric(1)}]\cr
Probability of node mutation. Note that each node is subject to mutation
independent of the the nodes.
Default is \eqn{0.1}, i.e., in expectation 10\% of the points are mutated.}

\item{...}{[any]\cr
Currently not used.}
}
\value{
[\code{matrix}] Mutated coordinates.
}
\description{
A subset \eqn{Q \subseteq P} of the points is selected and rotated
by a randomly sampled angle around its center.
}
\seealso{
\code{\link{build}}

Other mutation operators: \code{\link{doAxisProjectionMutation}},
  \code{\link{doClusterMutation}},
  \code{\link{doCompressionMutation}},
  \code{\link{doExpansionMutation}},
  \code{\link{doExplosionMutation}},
  \code{\link{doGridMutation}},
  \code{\link{doImplosionMutation}},
  \code{\link{doLinearProjectionMutation}},
  \code{\link{doNormalMutation}},
  \code{\link{doUniformMutation}}
}
\concept{mutation operators}
