% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{init}
\alias{init}
\title{Initialize a bare mutation operator collection.}
\usage{
init()
}
\value{
[\code{tspgen_collection}] Collection of mutation operators.
}
\description{
Initialize a bare mutation operator collection.
}
