% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{setProbabilities}
\alias{setProbabilities}
\title{Set mutation operator probabilities.}
\usage{
setProbabilities(collection, probs)
}
\arguments{
\item{collection}{[\code{tspgen_collection}]\cr
Collection of mutation operators.}

\item{probs}{[\code{numeric}]\cr
Numeric vector of probabilities. Needs to have the length of the mutation operator
collection, i.e., the number of mutation operators in the collection.}
}
\value{
[\code{tspgen_collection}] Augmented mutation operator collection.
}
\description{
This function defines the mutation operator probabilities. I.e.,
a vector of probabilities with the i-th component being the probability of
applying the i-th mutation operator in a single step of the instance generation
process.
}
